+++ test/nodes/e_base_records_test.rb
@@ -76,4 +76,40 @@ class EBaseRecordsTest < Minitest::Test
     assert_equal({ Canql::EQUALS => ['PAEDIATRIC'] },
                  parser.meta_data['unprocessed_records.sources'])
   end
+
+  def test_should_filter_by_unprocessed_enote2_records
+    parser = Canql::Parser.new('all cases with unprocessed enote records')
+    assert parser.valid?
+    assert_equal({ Canql::EQUALS => ['ENOTE2'] },
+                 parser.meta_data['unprocessed_records.sources'])
+
+    parser = Canql::Parser.new('all cases with unprocessed electronic notification records')
+    assert parser.valid?
+    assert_equal({ Canql::EQUALS => ['ENOTE2'] },
+                 parser.meta_data['unprocessed_records.sources'])
+  end
+
+  def test_should_filter_by_unprocessed_badger_records
+    parser = Canql::Parser.new('all cases with unprocessed badger records')
+    assert parser.valid?
+    assert_equal({ Canql::EQUALS => ['BADGER'] },
+                 parser.meta_data['unprocessed_records.sources'])
+  end
+
+  def test_should_filter_by_unprocessed_ucyto_records
+    parser = Canql::Parser.new('all cases with unprocessed ucyto records')
+    assert parser.valid?
+    assert_equal({ Canql::EQUALS => ['UCYTO'] },
+                 parser.meta_data['unprocessed_records.sources'])
+
+    parser = Canql::Parser.new('all cases with unprocessed cyto records')
+    assert parser.valid?
+    assert_equal({ Canql::EQUALS => ['UCYTO'] },
+                 parser.meta_data['unprocessed_records.sources'])
+
+    parser = Canql::Parser.new('all cases with unprocessed cytogenetic records')
+    assert parser.valid?
+    assert_equal({ Canql::EQUALS => ['UCYTO'] },
+                 parser.meta_data['unprocessed_records.sources'])
+  end
 end